refactor: move detailed card to pages, make more responsible design
