+++ src/features/SearchForm/ui/SearchForm.module.css
@@ -25,6 +25,10 @@
   opacity: 1;
 }
 
+.clear:disabled {
+  pointer-events: none;
+}
+
 @media (max-width: 768px) {
   .form {
     gap: 0.5em;
