+++ src/pages/DetailedCardPage/index.ts
