+++ src/pages/DetailedCardPage/ui/DetailedCard.module.css
@@ -2,6 +2,6 @@
   display: flex;
   justify-content: center;
   align-self: flex-end;
-  max-width: 30px;
-  min-height: 30px;
+  min-width: 45px;
+  min-height: 45px;
 }
