+++ src/pages/DetailedCardPage/ui/DetailedCard.test.tsx
