+++ src/pages/DetailedCardPage/ui/DetailedCard.tsx
