+++ src/pages/MainPage/ui/MainPage.module.css
@@ -43,5 +43,22 @@
   flex-direction: column;
   gap: 1em;
   height: calc(100vh - 167px - 100px);
+  padding-right: 0.3em;
   overflow-y: auto;
+}
+
+@media (max-width: 768px) {
+  .section.detailed {
+    display: flex;
+    flex-direction: column-reverse;
+  }
+
+  .right-column {
+    position: relative;
+    top: 0;
+    display: flex;
+    flex-direction: column;
+    gap: 1em;
+    height: 100%;
+  }
 }
